feat(theme provider): renames component source file PE-4360
